Write config.ini with quotes - fixes #259
This also updates the installer to save most
config values directly to the database, instead
of relying on Config::importAll to import the
values from config.ini and rewrite the file on
the first run, which never really made sense.

Eventually I plan to implement #215 by replacing
config.ini with a config.php file, which will
work around issues like this more permanently.