New backlog UI finished, now for the Taskboard.
This includes minor changes to how the dashboard widgets for types are
displayed. Previously, each widget, e.g. Projects, would filter to it's
type by the issue_types.project config value. Now, the type roles are
used to load all types with the Project role. This provides behavior
consistent with the new backlog.