Allow streaming video files, sort of.
This ideally should be expanded to allow reading attachments from
specific start positions, further expanding what's possible with
streaming large video files directly. For now, this will send the entire
video as a single response, and the browser can choose to play it if
that functionality is supported for the container and codec.